requirements.txt
absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
idna==3.2
ImageHash==4.2.1
imageio==2.9.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
tifffile==2021.8.30
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
